added new use case
GetCharacterUseCase.kt, gets Character Details by id
