+++ data/src/main/java/com/example/data/repository/CharacterRemoteRepositoryImpl.kt
@@ -1,6 +1,7 @@
 package com.example.data.repository
 
 import com.example.data.retrofit.RickAndMortyService
+import com.example.domain.model.CharacterDetails
 import com.example.domain.model.Characters
 import com.example.domain.repository.CharacterRemoteRepository
 import javax.inject.Inject
@@ -19,4 +20,11 @@ class CharacterRemoteRepositoryImpl @Inject constructor(
             rickAndMortyService.getCharacters(page, species, searchBy)
         }
     }
+
+    override suspend fun getCharacter(id: Int): Result<CharacterDetails> {
+        return runCatching {
+            rickAndMortyService.getCharacter(id)
+        }
+    }
+
 }
