+++ data/src/main/java/com/example/data/retrofit/RickAndMortyService.kt
@@ -1,7 +1,9 @@
 package com.example.data.retrofit
 
+import com.example.domain.model.CharacterDetails
 import com.example.domain.model.Characters
 import retrofit2.http.GET
+import retrofit2.http.Path
 import retrofit2.http.Query
 
 interface RickAndMortyService {
@@ -12,4 +14,9 @@ interface RickAndMortyService {
         @Query("species") species: String,
         @Query("name") searchBy: String,
     ): Characters
+
+    @GET("character/{id}")
+    suspend fun getCharacter(
+        @Path("id") id: Int
+    ): CharacterDetails
 }
