+++ domain/src/main/java/com/example/domain/model/CharacterDetails.kt
@@ -1,0 +1,24 @@
+package com.example.domain.model
+
+data class CharacterDetails(
+    val id: Int,
+    val name: String,
+    val status: String,
+    val species: String,
+    val type: String,
+    val gender: String,
+    val origin: Origin,
+    val location: Location,
+    val image: String,
+    val created: String
+)
+
+data class Origin(
+    val name: String,
+    val url: String
+)
+
+data class Location(
+    val name: String,
+    val url: String
+)
