+++ domain/src/main/java/com/example/domain/repository/CharacterRemoteRepository.kt
@@ -1,9 +1,14 @@
 package com.example.domain.repository
 
+import com.example.domain.model.CharacterDetails
 import com.example.domain.model.Characters
 
 interface CharacterRemoteRepository {
 
-    suspend fun getCharacters(page: Int, species: String = "", searchBy: String = ""): Result<Characters>
+    suspend fun getCharacters(
+        page: Int, species: String = "", searchBy: String = ""
+    ): Result<Characters>
+
+    suspend fun getCharacter(id: Int): Result<CharacterDetails>
 
 }
